md_hash field in g_eli_metadata structure is not 4 byte aligned, which case panic on sparc64.
The problem is in MD5(9) implementation. The Encode() function takes
'unsigned char *output' as its first argument, which is then assigned to
'u_int32_t *op'. If the 'output' argument is not 4 byte aligned (and in
geli(8) case it is not), sparc64 machine will panic.

I don't know how to fix MD5(9) in a clean way, so I'm implementing a
work-around in geli(8).

Reported by:	brueffer
MFC after:	3 days


git-svn-id: svn://svn.freebsd.org/base/head@157293 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f